refactor(storage): extract optimization/strategy/tenant leaf crates (decomp slice B)
Batch B of the root-crate decomposition (plan:
docs/12-design/ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.adoc). Extracts three
cold, clean leaf subtrees out of the monolithic root crate into workspace crates,
batched into one PR to amortize a single CI run (per the PR-sizing mandate):

- proximadb-storage-optimization  <- src/storage/optimization (~486 LOC)
- proximadb-storage-strategy      <- src/storage/strategy (~314 LOC)
- proximadb-storage-tenant        <- src/storage/tenant (~3.6K LOC)

Each is a pure git-mv + re-export: src/storage/mod.rs now does
`pub use proximadb_storage_<name> as <name>`, so all existing call sites
(`crate::storage::{optimization,strategy,tenant}::*`, incl. tenant's 153 refs)
resolve unchanged. The only in-file edits are mechanical path repoints:
`crate::proto::` -> `proximadb_proto::`, and tenant's 16 `crate::storage::tenant::`
self-refs -> `crate::`.

No behavioural change. Verified: the three crates build; full
`cargo check -p proximadb --lib` green. Root crate src/ shrinks 973,469 -> 969,809
lines (kv slice A + this batch). Next clean leaves (encryption/schema) need their
intra-storage deps (persistence) inverted first.
