+++ crates/storage/proximadb-storage-optimization/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "proximadb-storage-optimization"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB storage-layer optimization helpers (metadata sorter)"
+
+[lib]
+name = "proximadb_storage_optimization"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+tracing.workspace = true
+proximadb-data-model.workspace = true
+proximadb-records.workspace = true
+proximadb-proto.workspace = true
