+++ crates/storage/proximadb-storage-optimization/src/lib.rs
