+++ crates/storage/proximadb-storage-strategy/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "proximadb-storage-strategy"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB storage-layout strategy types"
+
+[lib]
+name = "proximadb_storage_strategy"
+path = "src/lib.rs"
+
+[dependencies]
+serde_json.workspace = true
+proximadb-proto.workspace = true
