+++ crates/storage/proximadb-storage-strategy/src/lib.rs
@@ -12,7 +12,7 @@
 use std::collections::HashMap;
 
 // Use proto enums as base types for consistency
-use crate::proto::proximadb_v1::{
+use proximadb_proto::proximadb_v1::{
     DistanceMetric as ProtoDistanceMetric, IndexingAlgorithm as ProtoIndexingAlgorithm,
     StorageEngine as ProtoStorageEngine,
 };
