+++ crates/storage/proximadb-storage-tenant/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "proximadb-storage-tenant"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB storage-layer tenant model (context, domain, RBAC, resources, entity/knowledge-graph stores)"
+
+[lib]
+name = "proximadb_storage_tenant"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+dashmap.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+uuid = { version = "1", features = ["v4"] }
+proximadb-proto.workspace = true
+proximadb-storage-common.workspace = true
